make python client a standalone module and re-export
